Fix: 마이페이지 회원 탈퇴 코드 수정
회원 탈퇴 요청 시 패스워드 확인 및 로그아웃도 되도록 수정

+++ src/main/java/com/ssafy/enjoytrip/user/service/UserServiceImpl.java
@@ -67,7 +67,14 @@ public class UserServiceImpl implements UserService {
 
     @Transactional
     @Override
-    public int dropOutById(int id) {
+    public int dropOutById(int id, String password) {
+        // 기존 유저의 비밀번호와 회원정보 수정을 위해 입력한 비밀번호가 일치하는지 확인
+        String originalPassword = userDao.findById(id).getPassword();
+        // 일치하지 않을 경우, 수정 불가 에러 리턴
+        if (!passwordCheck(password, originalPassword)) {
+            throw new PasswordFailException();
+        }
+        // 일치할 경우
         return userDao.deleteById(id);
     }
